requirements.txt
pydantic-settings
requests
google-auth
google-adk
google-cloud-aiplatform[adk,agent-engines]
tabulate
llama-index

fastapi        
python-dotenv
uvicorn
